backend/app/schemas/user.py
from datetime import datetime

from pydantic import BaseModel, ConfigDict


class UserCreate(BaseModel):
    name: str
    phone_number: str


class UserRead(BaseModel):
    id: int
    name: str
    phone_number: str
    created_at: datetime

    model_config = ConfigDict(from_attributes=True)
