backend/pyproject.toml
[project]
name = "dinner-planner-backend"
version = "0.1.0"
description = "Dinner Planner Phase 2A backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "sqlalchemy>=2.0.38",
  "psycopg[binary]>=3.2.6",
  "alembic>=1.15.2",
  "python-dateutil>=2.9.0",
  "python-multipart>=0.0.20"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "anyio>=4.8.0"
]

[build-system]
requires = ["setuptools>=75.8.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["app"]
testpaths = ["tests"]
